Add feature where users can only add comments when logged in and refactor code
