+++ portfolio/src/main/webapp/index.html
@@ -7,6 +7,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@500&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@500&family=Roboto+Slab&display=swap" rel="stylesheet">
     <script src="script.js"></script>
+    <script src="comments.js"></script>
   </head>
   <body>
     <div id="nav-bar">
